moved the math stuff to a separated helper
