+++ src/main/java/teamport/aether/helper/AetherMathHelper.java
@@ -23,4 +23,11 @@ public class AetherMathHelper {
     public static double nextExponential(Random random) {
         return -Math.log(1 - random.nextDouble());
     }
+
+    public static double distanceToSqr(int x, int y, int z, int x1, int y1, int z1) {
+        double d3 = x - x1;
+        double d4 = y - y1;
+        double d5 = z - z1;
+        return d3 * d3 + d4 * d4 + d5 * d5;
+    }
 }
