+++ src/main/java/teamport/aether/world/generate/feature/WorldFeatureAetherDungeonGold.java
@@ -13,6 +13,7 @@ import net.minecraft.core.world.generate.feature.WorldFeatureFlowers;
 import net.minecraft.core.world.generate.feature.WorldFeatureTallGrass;
 import teamport.aether.blocks.AetherBlocks;
 import teamport.aether.entity.boss.sunspirit.MobBossSunspirit;
+import teamport.aether.helper.AetherMathHelper;
 import teamport.aether.world.DungeonMapEntry;
 import teamport.aether.world.generate.feature.components.WorldFeatureComponent;
 import teamport.aether.world.generate.feature.components.WorldFeaturePoint;
@@ -262,7 +263,7 @@ public class WorldFeatureAetherDungeonGold extends WorldFeature{
     public void decorateTopLevelInRadius(Pair<Integer, WorldFeature>[] worldFeaturePair, int radius, int x, int y, int z) {
         int radX, radZ, height;
         for (radX = -radius; radX < radius; radX++) for (radZ = -radius; radZ < radius; radZ++) {
-            if (WorldFeatureComponent.distanceToSqr((radX + x), y, (radZ + z), x, y, z) < Math.pow(radius, 2)) {
+            if (AetherMathHelper.distanceToSqr((radX + x), y, (radZ + z), x, y, z) < Math.pow(radius, 2)) {
                 WorldFeatureComponent decorator = new WorldFeatureComponent();
                 height = world.getHeightValue((radX + x), (radZ + z));
                 if (Math.abs(height - y) > radius*2.25) continue;
